Add: agents
Agents are now generated and stored in starmap.xml.  They will also be
displayed in planet detail page, though right now the display is full of
placeholders in the form of dummy text and icoGenAgent for picture.

Need to create an appropriate function to translate agent ID (a unique 3
digit number) into an agent type by grabbing the agent ID from
agents.xml and returning the stored type.